+++ models/users.js
@@ -32,7 +32,7 @@ const UserSchema            = new mongoose.Schema({
         firstName           : { type: String, required: true },
         lastName            : { type: String, required: true },
         gender              : { type: String, required: true, enum: ['Male', 'Female', 'NA'], default: 'NA' },
-        residence           : { type: String },
+        residency           : { type: String },
         timezone            : { type: String },
         birthday            : { type: Date },
     },